refactoring

+++ src/azure_utils.py
@@ -8,7 +8,8 @@ from app_auth import get_access_token
 logging.basicConfig(level=logging.ERROR)
 
 # TODO: da aggiungere in folder utils
-def update_header(access_token):
+def get_http_header(access_token):
+
     return{
         'Authorization': f'Bearer {access_token}',
         'Content-type': 'application/json',
@@ -18,11 +19,77 @@ def update_header(access_token):
 def is_user_in_group():
     return
 
+def get_group_by_name(group_name, access_token):
+
+    url = f'{config.GRAPH_BASE_URL}/groups/'
+
+    params = {
+        '$count': 'true',
+        '$search': f'"displayName:{group_name}"',
+        '$select': 'displayName,id'
+    }
+
+    header = get_http_header(access_token)
+
+    try:
+        respone_group_info = requests.get(url,headers=header, params=params)
+        respone_group_info.raise_for_status()
+        groups_data = respone_group_info.json()
+
+        # Nessun gruppo trovato
+        if groups_data['@odata.count'] == 0:  
+            return {
+                'status_code':respone_group_info.status_code,
+                'error':f'No AAD group that contains {group_name} found. Try another name.'
+            }
+        
+        else:
+            count_match_name = 0
+            data = {}
+            for group in groups_data['value']:
+                if group_name in group['displayName']:
+                    count_match_name = count_match_name + 1
+                    data = {
+                        'status_code':respone_group_info.status_code,
+                        'group_id':group['id'],
+                        'group_name':group['displayName']
+                    }
+            # Diversi gruppi che contengono la substring
+            if count_match_name > 1:
+                return {
+                    'status_code': respone_group_info.status_code,
+                    'error':f'Too much AAD group that contains {group_name} found. Try another name.'
+                }
+            
+            # Singolo gruppo AAD trovato
+            elif count_match_name == 1:
+                return data
+            else:
+                return {
+                'status_code': respone_group_info.status_code,
+                'error':f'No AAD group that contains {group_name} found. Try another name.'
+            }
+
+
+    except requests.exceptions.HTTPError as e:
+        error_message = e.response.json().get('error', {}).get('message',str(e))
+        logging.error(f'Error: {error_message}.')
+        return {
+            'status_code': e.response.status_code,
+            'error':error_message
+        }
+    except requests.exceptions.RequestException as e:
+        logging.error(f'Request failed: {e}')
+        return {
+            'status_code': None,
+            'error': str(e)
+        } 
+
 def get_user_from_upn (UPN, access_token ):
 
     url = f'{config.GRAPH_BASE_URL}/users/{UPN}' 
 
-    headers = update_header(access_token)
+    headers = get_http_header(access_token)
 
     try:
         response = requests.get(url, headers=headers)
@@ -63,7 +130,7 @@ def get_user_group_by_name (user_id,group_name,access_token):
         '$select': 'displayName,id'
     }
 
-    header = update_header(access_token)
+    header = get_http_header(access_token)
 
     try:
         result = requests.get(url,headers=header, params=params)
@@ -95,7 +162,25 @@ def get_user_group_by_name (user_id,group_name,access_token):
             'error': str(e)
         }
     
-def add_user_to_group(user_id, group_id, access_token):
+def add_user_to_group(user_upn, group_name, access_token):
+
+    # Get user Id from UPN
+    response_user_info = get_user_from_upn(user_upn, access_token)
+
+    if 'error' in response_user_info:
+        return response_user_info    
+    user_id = response_user_info.get('id')
+
+    # Find AAD group by name   
+    response_user_group = get_group_by_name(group_name,access_token)
+
+    if 'error' in response_user_group:
+        return response_user_group
+    
+    group_id = response_user_group.get('group_id')
+    group_name = response_user_group.get('group_name')
+
+
 
     url = f'{config.GRAPH_BASE_URL}/groups/{group_id}/members/$ref'
 
@@ -103,14 +188,14 @@ def add_user_to_group(user_id, group_id, access_token):
         '@odata.id': f'{config.GRAPH_BASE_URL}/directoryObjects/{user_id}'
     }
 
-    headers = update_header(access_token)
+    headers = get_http_header(access_token)
     try:
         response = requests.post(url, headers=headers, json=payload)
         response.raise_for_status()
 
         return {
             'status_code':response.status_code,
-            'message': f'Success. User {user_id} added to AAD group {group_id}.'
+            'message': f'Success. User {user_upn} added to AAD group {group_name}.'
         }
     
     except requests.exceptions.HTTPError as e:
@@ -127,10 +212,25 @@ def add_user_to_group(user_id, group_id, access_token):
             'error': str(e)
         }
 
-def remove_user_from_group(user_id, group_id, access_token):
+def remove_user_from_group(user_upn, group_name, access_token):
+
+    response_user_info = get_user_from_upn(user_upn, access_token)
+
+    if 'error' in response_user_info:
+        return response_user_info    
+    user_id = response_user_info.get('id')
+
+    # Find user membership by group name    
+    response_user_group = get_user_group_by_name(user_id,group_name,access_token)
+
+    if 'error' in response_user_group:
+        return response_user_group
+    
+    group_id = response_user_group.get('group_id')
+    group_name = response_user_group.get('group_name')
 
     url = f'{config.GRAPH_BASE_URL}/groups/{group_id}/members/{user_id}/$ref'
-    headers = update_header(access_token)
+    headers = get_http_header(access_token)
     try:
         
         response = requests.delete(url, headers=headers)
@@ -138,7 +238,7 @@ def remove_user_from_group(user_id, group_id, access_token):
 
         return {
             'status_code':response.status_code,
-            'message': f'Success. User {user_id} removed from AAD group {group_id}.'
+            'message': f'Success. User {user_upn} removed from AAD group {group_name}.'
         }
 
     except requests.exceptions.HTTPError as e:
